# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "description": "Api for all the function",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/recipes": {
            "get": {
                "tags": [
                    "Recipes"
                ],
                "summary": "Get all recipes",
                "description": "Returns a list of all recipes.",
                "operationId": "getRecipesList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "New Recipe"
                                            },
                                            "chef_id": {
                                                "type": "interger",
                                                "example": "1"
                                            },
                                            "unique_code": {
                                                "type": "string",
                                                "example": "UNI123"
                                            },
                                            "popularity": {
                                                "type": "interger",
                                                "example": "10"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2024-06-28 12:00:00"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2024-06-28 12:00:00"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Recipes"
                ],
                "summary": "Create a new recipe",
                "description": "Creates a new recipe in the database.",
                "operationId": "createRecipe",
                "requestBody": {
                    "description": "Recipe object to be created",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "chef_id",
                                    "unqiue_code"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "New recipe"
                                    },
                                    "chef_id": {
                                        "type": "interger",
                                        "example": "1"
                                    },
                                    "unique_code": {
                                        "type": "string",
                                        "example": "UNI234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Recipe created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "New Recipe"
                                        },
                                        "chef_id": {
                                            "type": "interger",
                                            "example": "1"
                                        },
                                        "unique_code": {
                                            "type": "string",
                                            "example": "UNI123"
                                        },
                                        "popularity": {
                                            "type": "interger",
                                            "example": "10"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-06-28 12:00:00"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-06-28 12:00:00"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/recipes/{recipe}": {
            "get": {
                "tags": [
                    "Recipes"
                ],
                "summary": "Retrieve a recipe by ID",
                "description": "Returns a single recipe based on its ID.",
                "operationId": "getRecipeById",
                "parameters": [
                    {
                        "name": "recipe",
                        "in": "path",
                        "description": "ID of recipe to fetch",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recipe retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "New Recipe"
                                        },
                                        "chef_id": {
                                            "type": "interger",
                                            "example": "1"
                                        },
                                        "unique_code": {
                                            "type": "string",
                                            "example": "UNI123"
                                        },
                                        "popularity": {
                                            "type": "interger",
                                            "example": "10"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-06-28 12:00:00"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-06-28 12:00:00"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Recipe not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Recipes"
                ],
                "summary": "Update an existing recipe",
                "description": "Updates a recipe based on its ID.",
                "operationId": "updateRecipe",
                "parameters": [
                    {
                        "name": "recipe",
                        "in": "path",
                        "description": "ID of recipe to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Recipe object to be created",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "chef_id",
                                    "unqiue_code"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated recipe"
                                    },
                                    "chef_id": {
                                        "type": "interger",
                                        "example": "1"
                                    },
                                    "unique_code": {
                                        "type": "string",
                                        "example": "UNI234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Recipe created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Updated] Recipe"
                                        },
                                        "chef_id": {
                                            "type": "interger",
                                            "example": "1"
                                        },
                                        "unique_code": {
                                            "type": "string",
                                            "example": "UNI123"
                                        },
                                        "popularity": {
                                            "type": "interger",
                                            "example": "10"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-06-28 12:00:00"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-06-28 12:00:00"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Recipe not found"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Recipes"
                ],
                "summary": "Delete a recipe",
                "description": "Deletes a recipe based on its ID.",
                "operationId": "deleteRecipe",
                "parameters": [
                    {
                        "name": "recipe",
                        "in": "path",
                        "description": "ID of recipe to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Recipe successfully deleted"
                    },
                    "404": {
                        "description": "Recipe not found"
                    }
                }
            }
        },
        "/api/recipes/{recipe}/save-ingredients": {
            "post": {
                "tags": [
                    "Recipes"
                ],
                "summary": "Save ingredients for a recipe",
                "description": "Associates ingredients with a recipe based on recipe ID.",
                "operationId": "saveRecipeIngredients",
                "parameters": [
                    {
                        "name": "recipe",
                        "in": "path",
                        "description": "ID of recipe to associate ingredients with",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "RecipeIngredient object containing ingredient IDs",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "ingredient_id"
                                ],
                                "properties": {
                                    "ingredient_id": {
                                        "description": "Array of ingredient IDs",
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successfully saved recipe ingredients",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully saved recipe ingredients."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        }
    }
}